+++ types/function/memoize.d.ts
@@ -1,0 +1,1 @@
+export { memoize, memoize as default } from "../function.js";