+++ types/function/once.d.ts
@@ -1,0 +1,1 @@
+export { once, once as default } from "../function.js";